pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "open3e-bridge"
version = "0.1.0"
description = "Open3E Home Assistant MQTT Discovery Bridge"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
dependencies = [
    "paho-mqtt>=2.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
    "jinja2",
]
e2e = [
    "docker>=7.0",
]

[tool.setuptools]
# Include bridge.py as top-level module (alongside generators/ package)
py-modules = ["bridge"]

[tool.setuptools.packages.find]
include = ["generators*", "config*"]

[tool.setuptools.package-data]
config = ["**/*.yaml"]

[project.scripts]
open3e-bridge = "bridge:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["e2e: End-to-end tests (require Docker)"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
